build from source in Dockerfile
In the Dockerfile, the GitHub release asset were downloaded and stored in the Docker image.
The release asset generation process is triggered by putting a git tag.
When Docker image generation process is also triggered by putting a git tag, there is no GitHub Release asset for that version.

Therefore, I changed Dockerfile to build from source so that it does not depend on the version of the GitHub Release asset.